refactor(test-runner): tidy things up
